move XmlError + Deserializer into separate module
